requirements.txt
discord.py
aiohttp[speedups]
aiomysql
toml
uvloop